# Preserve caches on context switch