--- a/AUTHORS
+++ b/AUTHORS
@@ -10,6 +10,7 @@ Other contributors, listed alphabetically, are:
 * Brend Wanders <b.wanders@utwente.nl>
 * choloepus
 * coutinho <coutinho@esrf.fr>
+* Clément Pit-Claudel <clement.pitclaudel@live.com>
 * Daniel Sokolowski <daniel.sokolowski@danols.com>
 * Dave Brooks <dave@bcs.co.nz>
 * David Linke