--- a/AUTHORS
+++ b/AUTHORS
@@ -20,6 +20,7 @@ Other contributors, listed alphabetically, are:
 * Felix Hummel <felix@felixhummel.de>
 * Francisco Couzo <franciscouzo@gmail.com>
 * Giel van Schijndel <me@mortis.eu>
+* Guido Imperiale <crusaderky@gmail.com>
 * Ignacio Fdez. Galván <jellby@yahoo.com>
 * James Rowe <jnrowe@gmail.com>
 * Jim Turner <jturner314@gmail.com>