--- a/pint/registry.py
+++ b/pint/registry.py
@@ -75,6 +75,21 @@ class _Meta(type):
         obj._after_init()
         return obj
 
+class RegistryCache():
+    """Cache to speed up unit registries"""
+
+    def __init__(self):
+        #: Maps dimensionality (UnitsContainer) to Units (str)
+        self.dimensional_equivalents = dict()
+
+        #: Maps dimensionality (UnitsContainer) to Dimensionality (UnitsContainer)
+        self.root_units = dict()
+
+        #: Maps dimensionality (UnitsContainer) to Units (UnitsContainer)
+        self.dimensionality = dict()
+
+        #: Cache the unit name associated to user input. ('mV' -> 'millivolt')
+        self.parse_unit = dict()
 
 class BaseRegistry(meta.with_metaclass(_Meta)):
     """Base class for all registries.
@@ -154,17 +169,8 @@ class BaseRegistry(meta.with_metaclass(_Meta)):
         #: Map suffix name (string) to canonical , and unit alias to canonical unit name
         self._suffixes = {'': None, 's': ''}
 
-        #: Maps dimensionality (UnitsContainer) to Units (str)
-        self._dimensional_equivalents = dict()
-
-        #: Maps dimensionality (UnitsContainer) to Dimensionality (UnitsContainer)
-        self._root_units_cache = dict()
-
-        #: Maps dimensionality (UnitsContainer) to Units (UnitsContainer)
-        self._dimensionality_cache = dict()
-
-        #: Cache the unit name associated to user input. ('mV' -> 'millivolt')
-        self._parse_unit_cache = dict()
+        #: Map contexts to RegistryCache
+        self._cache = RegistryCache()
 
         self._initialized = False
 
@@ -404,7 +410,10 @@ class BaseRegistry(meta.with_metaclass(_Meta)):
     def _build_cache(self):
         """Build a cache of dimensionality and base units.
         """
-        self._dimensional_equivalents = dict()
+        self._cache = RegistryCache()
+        root_units = self._cache.root_units
+        dimensionality = self._cache.dimensionality
+        dimensional_equivalents = self._cache.dimensional_equivalents
 
         deps = dict((name, set(definition.reference.keys() if definition.reference else {}))
                     for name, definition in self._units.items())
@@ -426,14 +435,14 @@ class BaseRegistry(meta.with_metaclass(_Meta)):
                     bu = self._get_root_units(uc)
                     di = self._get_dimensionality(uc)
 
-                    self._root_units_cache[uc] = bu
-                    self._dimensionality_cache[uc] = di
+                    root_units[uc] = bu
+                    dimensionality[uc] = di
 
                     if not prefixed:
-                        if di not in self._dimensional_equivalents:
-                            self._dimensional_equivalents[di] = set()
+                        if di not in dimensional_equivalents:
+                            dimensional_equivalents[di] = set()
 
-                        self._dimensional_equivalents[di].add(self._units[base_name]._name)
+                        dimensional_equivalents[di].add(self._units[base_name]._name)
 
                 except Exception as e:
                     logger.warning('Could not resolve {0}: {1!r}'.format(unit_name, e))
@@ -528,8 +537,10 @@ class BaseRegistry(meta.with_metaclass(_Meta)):
         if not input_units:
             return UnitsContainer()
 
-        if input_units in self._dimensionality_cache:
-            return self._dimensionality_cache[input_units]
+        cache = self._cache.dimensionality
+
+        if input_units in cache:
+            return cache[input_units]
 
         accumulator = defaultdict(float)
         self._get_dimensionality_recurse(input_units, 1.0, accumulator)
@@ -540,7 +551,7 @@ class BaseRegistry(meta.with_metaclass(_Meta)):
         dims = UnitsContainer(dict((k, v) for k, v in accumulator.items()
                                    if v != 0.0))
 
-        self._dimensionality_cache[input_units] = dims
+        cache[input_units] = dims
 
         return dims
 
@@ -615,9 +626,11 @@ class BaseRegistry(meta.with_metaclass(_Meta)):
         if not input_units:
             return 1., UnitsContainer()
 
+        cache = self._cache.root_units
+
         # The cache is only done for check_nonmult=True
-        if check_nonmult and input_units in self._root_units_cache:
-            return self._root_units_cache[input_units]
+        if check_nonmult and input_units in cache:
+            return cache[input_units]
 
         accumulators = [1., defaultdict(float)]
         self._get_root_units_recurse(input_units, 1.0, accumulators)
@@ -633,7 +646,7 @@ class BaseRegistry(meta.with_metaclass(_Meta)):
                     return None, units
 
         if check_nonmult:
-            self._root_units_cache[input_units] = factor, units
+            cache[input_units] = factor, units
 
         return factor, units
 
@@ -683,7 +696,7 @@ class BaseRegistry(meta.with_metaclass(_Meta)):
 
         src_dim = self._get_dimensionality(input_units)
 
-        ret = self._dimensional_equivalents[src_dim]
+        ret = self._cache.dimensional_equivalents[src_dim]
 
         return ret
 
@@ -795,8 +808,10 @@ class BaseRegistry(meta.with_metaclass(_Meta)):
         if as_delta is None:
             as_delta = True
 
-        if as_delta and input_string in self._parse_unit_cache:
-            return self._parse_unit_cache[input_string]
+        cache = self._cache.parse_unit
+
+        if as_delta and input_string in cache:
+            return cache[input_string]
 
         if not input_string:
             return UnitsContainer()
@@ -824,7 +839,7 @@ class BaseRegistry(meta.with_metaclass(_Meta)):
         ret = UnitsContainer(ret)
 
         if as_delta:
-            self._parse_unit_cache[input_string] = ret
+            cache[input_string] = ret
 
         return ret
 
@@ -1042,6 +1057,9 @@ class ContextRegistry(BaseRegistry):
         #: Stores active contexts.
         self._active_ctx = ContextChain()
 
+        #: Map context chain to cache
+        self._caches = { self._active_ctx: self._cache }
+
     def _register_parsers(self):
         super(ContextRegistry, self)._register_parsers()
         self._register_parser('@context', self._parse_context)
@@ -1082,6 +1100,15 @@ class ContextRegistry(BaseRegistry):
             del self._contexts[alias]
 
         return context
+
+    def _build_cache(self):
+        """"""
+        cache = self._caches.get(self._active_ctx)
+        if cache is None:
+            super()._build_cache()
+            self._caches[self._active_ctx] = self._cache
+        else:
+            self._cache = cache
 
     def enable_contexts(self, *names_or_contexts, **kwargs):
         """Enable contexts provided by name or by object.
@@ -1251,7 +1278,7 @@ class ContextRegistry(BaseRegistry):
             nodes = find_connected_nodes(self._active_ctx.graph, src_dim)
             if nodes:
                 for node in nodes:
-                    ret |= self._dimensional_equivalents[node]
+                    ret |= self._cache.dimensional_equivalents[node]
 
         return ret