Closes #795 
Forked from #801