@crusaderky 2019-12-03
Benchmark:
```python
import pint
ureg = pint.UnitRegistry()
src = ureg('1 m')
dst = ureg.Unit('km')
%timeit src.to(dst)
%timeit -n 1 -r 1 src.to(dst, 'chemistry')
%timeit src.to(dst, 'chemistry')
ureg.enable_contexts('chemistry')
%timeit src.to(dst)
```
Before:
37.6 µs ± 921 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
78.9 ms ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)
83.4 ms ± 2.32 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
60.1 µs ± 1.37 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

After:
31.1 µs ± 2.57 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
107 ms ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)
85.1 µs ± 3.31 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
48.3 µs ± 1.92 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)