@crusaderky 2019-12-04
@hgrecco ready for review and merge